always copy the job artifacts to the storage directory

+++ train.sh
@@ -30,13 +30,18 @@ do
   esac
 done
 
+trap 'on_exit' EXIT
 
 python -m cProfile -o /artifacts/train.prof train.py $* 2>/artifacts/err.log | tee -a /artifacts/out.log
 
+function on_exit() {
+  if [ -f /artifacts/logs ]
+  then
+    ( cd /artifacts && zip -r logs.zip logs )
+    rm -rf /artifacts/logs
+  fi
 
-( cd /artifacts && zip -q -r logs.zip logs )
-rm -rf /artifacts/logs
-
-rm -rf /storage/models/tgs/${RUN_NAME}
-mkdir -p /storage/models/tgs/${RUN_NAME}
-cp -r /artifacts/* /storage/models/tgs/${RUN_NAME}
+  rm -rf /storage/models/tgs/${RUN_NAME}
+  mkdir -p /storage/models/tgs/${RUN_NAME}
+  cp -r /artifacts/* /storage/models/tgs/${RUN_NAME}
+}
